+++ MQTTClient/MQTTClient/MQTTSessionManager.h
@@ -27,8 +27,6 @@ typedef NS_ENUM(int, MQTTSessionManagerState) {
     MQTTSessionManagerStateClosed
 };
 
-
-
 /** gets called when a new message was received
  @param data the data received, might be zero length
  @param topic the topic the data was published to
@@ -60,9 +58,24 @@ typedef NS_ENUM(int, MQTTSessionManagerState) {
  */
 @property (nonatomic, readonly) NSError *lastErrorCode;
 
+/** initWithPersistence sets the MQTTPersistence properties other than default
+ * @param persistent YES or NO (default) to establish file or in memory persistence.
+ * @param maxWindowSize (a positive number, default is 16) to control the number of messages sent before waiting for acknowledgement in Qos 1 or 2. Additional messages are stored and transmitted later.
+ * @param maxSize (a positive number of bytes, default is 64 MB) to limit the size of the persistence file. Messages published after the limit is reached are dropped.
+ * @param maxMessages (a positive number, default is 1024) to limit the number of messages stored. Additional messages published are dropped.
+ * @return the initialized MQTTSessionManager object
+ */
+
+- (MQTTSessionManager *)initWithPersistence:(BOOL)persistent
+                              maxWindowSize:(NSUInteger)maxWindowSize
+                                maxMessages:(NSUInteger)maxMessages
+                                    maxSize:(NSUInteger)maxSize;
+
+
+
 /** Connects to the MQTT broker and stores the parameters for subsequent reconnects
  * @param host specifies the hostname or ip address to connect to. Defaults to @"localhost".
- * @param port spefies the port to connect to
+ * @param port specifies the port to connect to
  * @param tls specifies whether to use SSL or not
  * @param keepalive The Keep Alive is a time interval measured in seconds. The MQTTClient ensures that the interval between Control Packets being sent does not exceed the Keep Alive value. In the  absence of sending any other Control Packets, the Client sends a PINGREQ Packet.
  * @param clean specifies if the server should discard previous session information.
@@ -70,13 +83,48 @@ typedef NS_ENUM(int, MQTTSessionManagerState) {
  * @param user an NSString object containing the user's name (or ID) for authentication. May be nil.
  * @param pass an NSString object containing the user's password. If userName is nil, password must be nil as well.
  * @param will indicates whether a will shall be sent
- * @param willTopic the Will Topic is a string, must not be nil
- * @param willMsg the Will Message, might be zero length
+ * @param willTopic the Will Topic is a string, may be nil
+ * @param willMsg the Will Message, might be zero length or nil
  * @param willQos specifies the QoS level to be used when publishing the Will Message.
  * @param willRetainFlag indicates if the server should publish the Will Messages with retainFlag.
  * @param clientId The Client Identifier identifies the Client to the Server. If nil, a random clientId is generated.
- * @return the initialised MQTTSessionManager object
-*/
+ * @param securityPolicy A custom SSL security policy or nil.
+ * @param certificates An NSArray of the pinned certificates to use or nil.
+ */
+
+- (void)connectTo:(NSString *)host
+             port:(NSInteger)port
+              tls:(BOOL)tls
+        keepalive:(NSInteger)keepalive
+            clean:(BOOL)clean
+             auth:(BOOL)auth
+             user:(NSString *)user
+             pass:(NSString *)pass
+             will:(BOOL)will
+        willTopic:(NSString *)willTopic
+          willMsg:(NSData *)willMsg
+          willQos:(MQTTQosLevel)willQos
+   willRetainFlag:(BOOL)willRetainFlag
+     withClientId:(NSString *)clientId
+   securityPolicy:(MQTTSSLSecurityPolicy *)securityPolicy
+     certificates:(NSArray *)certificates;
+
+/** Convenience alternative to full paramter connectTo
+ * @param host see connectTo description
+ * @param port see connectTo description
+ * @param tls see connectTo description
+ * @param keepalive see connectTo description
+ * @param clean see connectTo description
+ * @param auth see connectTo description
+ * @param user see connectTo description
+ * @param pass see connectTo description
+ * @param will see connectTo description
+ * @param willTopic see connectTo description
+ * @param willMsg see connectTo description
+ * @param willQos see connectTo description
+ * @param willRetainFlag see connectTo description
+ * @param clientId see connectTo description
+ */
 
 - (void)connectTo:(NSString *)host
              port:(NSInteger)port
@@ -93,24 +141,23 @@ typedef NS_ENUM(int, MQTTSessionManagerState) {
    willRetainFlag:(BOOL)willRetainFlag
      withClientId:(NSString *)clientId;
 
-/** Connects to the MQTT broker and stores the parameters for subsequent reconnects
- * @param host specifies the hostname or ip address to connect to. Defaults to @"localhost".
- * @param port spefies the port to connect to
- * @param tls specifies whether to use SSL or not
- * @param keepalive The Keep Alive is a time interval measured in seconds. The MQTTClient ensures that the interval between Control Packets being sent does not exceed the Keep Alive value. In the  absence of sending any other Control Packets, the Client sends a PINGREQ Packet.
- * @param clean specifies if the server should discard previous session information.
- * @param auth specifies the user and pass parameters should be used for authenthication
- * @param user an NSString object containing the user's name (or ID) for authentication. May be nil.
- * @param pass an NSString object containing the user's password. If userName is nil, password must be nil as well.
+/** Convenience alternative to full paramter connectTo
+ * @param host see connectTo description
+ * @param port see connectTo description
+ * @param tls see connectTo description
+ * @param keepalive see connectTo description
+ * @param clean see connectTo description
+ * @param auth see connectTo description
+ * @param user see connectTo description
+ * @param pass see connectTo description
  * @param willTopic the Will Topic is a string, must not be nil
  * @param will the Will Message, might be zero length
- * @param willQos specifies the QoS level to be used when publishing the Will Message.
- * @param willRetainFlag indicates if the server should publish the Will Messages with retainFlag.
- * @param clientId The Client Identifier identifies the Client to the Server. If nil, a random clientId is generated.
- * @return the initialised MQTTSessionManager object
+ * @param willQos see connectTo description
+ * @param willRetainFlag see connectTo description
+ * @param clientId see connectTo description
  */
 
-     - (void)connectTo:(NSString *)host
+- (void)connectTo:(NSString *)host
                   port:(NSInteger)port
                    tls:(BOOL)tls
              keepalive:(NSInteger)keepalive